requirements.txt
flatland_rl==4.0.3
ipython==8.12.3
matplotlib==3.6.3
numpy==1.24.4
pandas==2.0.3
Pillow==11.0.0
pybind11==2.13.6
setuptools==65.5.0
torch==2.5.0
tqdm==4.66.6
yappi==1.6.10